Add files via upload
Use make_regression function in skearn to generate dataset with 500 samples with 3 features (x_1,x_2,x_3) with random_state=10. The hypothesis is defined as follows:
〖h(x)〗_θ=〖〖3θ〗_1 x〗_1^2+〖θ_2 x〗_2^3- θ_3 x_2^3 〖+ 〖〖θ_2 θ〗_3 x〗_3^3+θ〗_0

	Split the dataset into 60% train data and 40% test data. 

	Write a program to implement linear regression using gradient decent based. The loss function is defined as follows:
〖J(θ〗_0, θ_1,θ_2,θ_3)=1/2m ∑_(i=1)^m▒〖( h_θ (x^((i) )-y^((i) ) )^2 〗
	Plot the cost/loss function against the iterations (epochs)



Use make_regression function in skearn to generate dataset with 500 samples with 15 features with random_state=10. 
	Split the dataset into 70% train data and 30% test data. 
	Use sklearn to implement the following: 
	Linear regression 
	Lasso regression
	Ridge regression
	Polynomial linear regression

	Draw bar chart or table that compares the mean squared error between the correct outputs and the predicted outputs on testing set for linear models.
Q4) Normal equation (use dataset in Q3)
You learned normal equation (closed-form solution) that can be used for finding the optimal parameter values θ for linear regression given the matrix of training examples X and the corresponding response variables y. Write a Python code to implement normal equation. The formula is defined as follows:


Attached is the dataset and solution